+++ app/client/src/components/pages/Workspace/ActionBar.tsx
@@ -1,4 +1,4 @@
-import { SkipNext, SkipPrevious, Edit } from "@mui/icons-material";
+import { SkipNext, SkipPrevious } from "@mui/icons-material";
 import { CircularProgress } from "@mui/material";
 import Box from "@mui/material/Box";
 import Button from "@mui/material/Button";
@@ -17,11 +17,6 @@ interface IActionBarProps {
   onNextClick: () => void;
   onPrevClick: () => void;
   onRestartClick: () => void;
-  onAddExerciseClick: () => void;
-  onEditExerciseClick?: () => void;
-  onEditHintClick?: () => void;
-  isGitHubConnected?: boolean;
-  currentExerciseName?: string;
   isTest: boolean;
   succeeded: boolean;
   hintVisible: boolean;
@@ -36,11 +31,6 @@ export const ActionBar = ({
   onPrevClick,
   onNextClick,
   onRestartClick,
-  onAddExerciseClick,
-  onEditExerciseClick,
-  onEditHintClick,
-  isGitHubConnected,
-  currentExerciseName,
   isTest,
   succeeded,
   hintVisible,
@@ -49,7 +39,6 @@ export const ActionBar = ({
   last,
 }: IActionBarProps) => {
   const [dialogOpen, setDialogOpen] = useState(false);
-  const [editDialogOpen, setEditDialogOpen] = useState(false);
 
   const openDialog = () => {
     setDialogOpen(true);
@@ -57,19 +46,6 @@ export const ActionBar = ({
 
   const closeDialog = () => {
     setDialogOpen(false);
-  };
-
-  const openEditDialog = () => {
-    setEditDialogOpen(true);
-  };
-
-  const closeEditDialog = () => {
-    setEditDialogOpen(false);
-  };
-
-  const handleEditAction = (action: () => void) => {
-    closeEditDialog();
-    action();
   };
 
   return (
@@ -119,15 +95,6 @@ export const ActionBar = ({
             justifyContent: "flex-end",
           }}
         >
-          <Tooltip title="Edit">
-            <IconButton
-              onClick={openEditDialog}
-              sx={{ p: 0.5, color: "#FFF" }}
-              aria-label="edit-options"
-            >
-              <Edit />
-            </IconButton>
-          </Tooltip>
           <Button
             color="primary"
             variant="contained"
@@ -153,130 +120,6 @@ export const ActionBar = ({
           )}
         </Box>
       </Box>
-
-      <Dialog
-        open={editDialogOpen}
-        onClose={closeEditDialog}
-        aria-labelledby="edit-dialog-title"
-        maxWidth="sm"
-        fullWidth
-      >
-        <DialogTitle id="edit-dialog-title">
-          🛠️ Edit Options - {currentExerciseName}
-        </DialogTitle>
-        <DialogContent>
-          <DialogContentText sx={{ mb: 3 }}>
-            <strong>📋 Prerequisites:</strong>
-            <br />
-            • You must have the project <strong>forked</strong> in your GitHub account
-            <br />
-            • Your fork must be <strong>synced</strong> with the main repository
-            <br />
-            • You must be <strong>logged into GitHub</strong> in Starklings
-          </DialogContentText>
-          
-          <Box sx={{ display: "flex", flexDirection: "column", gap: 2 }}>
-            <Button
-              variant="contained"
-              onClick={() => handleEditAction(onAddExerciseClick)}
-              sx={{ 
-                justifyContent: "flex-start", 
-                textAlign: "left",
-                py: 1.5,
-                px: 2,
-                backgroundColor: "#2196f3",
-                "&:hover": {
-                  backgroundColor: "#1976d2"
-                }
-              }}
-            >
-              <Box>
-                <Box sx={{ fontWeight: "bold", mb: 0.5 }}>
-                  📚 View Contribution Guide
-                </Box>
-                <Box sx={{ fontSize: "0.85em", opacity: 0.9 }}>
-                  Learn how to add or edit exercises
-                </Box>
-              </Box>
-            </Button>
-
-            {isGitHubConnected && onEditExerciseClick && (
-              <Button
-                variant="contained"
-                onClick={() => handleEditAction(onEditExerciseClick)}
-                sx={{ 
-                  justifyContent: "flex-start", 
-                  textAlign: "left",
-                  py: 1.5,
-                  px: 2,
-                  backgroundColor: "#4caf50",
-                  "&:hover": {
-                    backgroundColor: "#388e3c"
-                  }
-                }}
-              >
-                <Box>
-                  <Box sx={{ fontWeight: "bold", mb: 0.5 }}>
-                    ✏️ Edit Exercise Code
-                  </Box>
-                  <Box sx={{ fontSize: "0.85em", opacity: 0.9 }}>
-                    Modify description and code of {currentExerciseName}
-                  </Box>
-                </Box>
-              </Button>
-            )}
-
-            {isGitHubConnected && onEditHintClick && (
-              <Button
-                variant="contained"
-                onClick={() => handleEditAction(onEditHintClick)}
-                sx={{ 
-                  justifyContent: "flex-start", 
-                  textAlign: "left",
-                  py: 1.5,
-                  px: 2,
-                  backgroundColor: "#ff9800",
-                  "&:hover": {
-                    backgroundColor: "#f57c00"
-                  }
-                }}
-              >
-                <Box>
-                  <Box sx={{ fontWeight: "bold", mb: 0.5 }}>
-                    💡 Edit Exercise Hint
-                  </Box>
-                  <Box sx={{ fontSize: "0.85em", opacity: 0.9 }}>
-                    Improve hints for {currentExerciseName} (goes to exact line)
-                  </Box>
-                </Box>
-              </Button>
-            )}
-
-            {!isGitHubConnected && (
-              <Box 
-                sx={{ 
-                  p: 2, 
-                  bgcolor: "rgba(255, 193, 7, 0.1)", 
-                  borderRadius: 1,
-                  border: "1px solid rgba(255, 193, 7, 0.3)"
-                }}
-              >
-                <Box sx={{ fontWeight: "bold", color: "#ff6f00", mb: 1 }}>
-                  ⚠️ GitHub not connected
-                </Box>
-                <Box sx={{ fontSize: "0.9em" }}>
-                  Connect your GitHub account to access exercise editing options.
-                </Box>
-              </Box>
-            )}
-          </Box>
-        </DialogContent>
-        <DialogActions>
-          <Button onClick={closeEditDialog} variant="contained">
-            Close
-          </Button>
-        </DialogActions>
-      </Dialog>
 
       {/* Restart Dialog (existing) */}
       <Dialog
